# NCHWc Upsample/Mul optimizations