--- a/onnxruntime/core/mlas/lib/snchwc.cpp
+++ b/onnxruntime/core/mlas/lib/snchwc.cpp
@@ -26,18 +26,18 @@ struct MLAS_NCHWC_WORK_BLOCK
     int32_t tids;
     size_t BatchCount;
     size_t InputChannels;
-    size_t InputShape[3];
+    size_t InputShape[2];
     size_t InputSize;
     size_t OutputChannels;
-    size_t OutputShape[3];
+    size_t OutputShape[2];
     size_t OutputSize;
-    size_t KernelShape[3];
-    size_t DilationShape[3];
-    size_t Padding[6];
-    size_t StrideShape[3];
-    size_t OutputCountLeftPad[3];
-    size_t OutputCount[3];
-    size_t OutputCountRightPad[3];
+    size_t KernelShape[2];
+    size_t DilationShape[2];
+    size_t Padding[4];
+    size_t StrideShape[2];
+    size_t OutputCountLeftPad[2];
+    size_t OutputCount[2];
+    size_t OutputCountRightPad[2];
 };
 
 //
@@ -111,7 +111,6 @@ Return Value:
 void
 MlasNchwcPrepareWorkBlock(
     MLAS_NCHWC_WORK_BLOCK* WorkBlock,
-    size_t Dimensions,
     const int64_t* InputShape,
     const int64_t* KernelShape,
     const int64_t* DilationShape,
@@ -130,8 +129,6 @@ Arguments:
 
     WorkBlock - Supplies the structure that contains the common convolution
         and pooling parameters.
-
-    Dimensions - Supplies the number of dimensions.
 
     InputShape - Supplies the shape of the input tensor.
 
@@ -169,9 +166,9 @@ Return Value:
 
     size_t InputSize = 1;
     size_t OutputSize = 1;
-    bool CanFlattenShape = (Dimensions == 2);
+    bool CanFlattenShape = true;
 
-    for (size_t dim = 0; dim < Dimensions; dim++) {
+    for (size_t dim = 0; dim < 2; dim++) {
 
         const size_t InputValue = size_t(InputShape[dim]);
         const size_t OutputValue = size_t(OutputShape[dim]);
@@ -198,13 +195,13 @@ Return Value:
 
         if (Padding != nullptr) {
             WorkBlock->Padding[dim] = size_t(Padding[dim]);
-            WorkBlock->Padding[dim + Dimensions] = size_t(Padding[dim + Dimensions]);
+            WorkBlock->Padding[dim + 2] = size_t(Padding[dim + 2]);
         } else {
             WorkBlock->Padding[dim] = 0;
-            WorkBlock->Padding[dim + Dimensions] = 0;
+            WorkBlock->Padding[dim + 2] = 0;
         }
 
-        CanFlattenShape &= (WorkBlock->Padding[dim] == 0 && WorkBlock->Padding[dim + Dimensions] == 0);
+        CanFlattenShape &= (WorkBlock->Padding[dim] == 0 && WorkBlock->Padding[dim + 2] == 0);
 
         if (StrideShape != nullptr) {
             WorkBlock->StrideShape[dim] = size_t(StrideShape[dim]);
@@ -249,7 +246,7 @@ Return Value:
     // Compute the number of output elements affected by left and right padding.
     //
 
-    for (size_t dim = 0; dim < Dimensions; dim++) {
+    for (size_t dim = 0; dim < 2; dim++) {
 
         const size_t SpanValue =
             WorkBlock->DilationShape[dim] * (WorkBlock->KernelShape[dim] - 1) + 1;
@@ -1214,7 +1211,6 @@ const PMLAS_POOL_FLOAT_KERNEL MLAS_NCHWC_POOL_ALGORITHM::PoolKernels[] =
 void
 MLASCALL
 MlasNchwcConv(
-    size_t Dimensions,
     const int64_t* InputShape,
     const int64_t* KernelShape,
     const int64_t* DilationShape,
@@ -1297,7 +1293,7 @@ Return Value:
     // Capture the generic shape parameters to the work block.
     //
 
-    MlasNchwcPrepareWorkBlock(&WorkBlock, Dimensions, InputShape, KernelShape,
+    MlasNchwcPrepareWorkBlock(&WorkBlock, InputShape, KernelShape,
         DilationShape, Padding, StrideShape, OutputShape);
 
     WorkBlock.InputChannels /= GroupCount;
@@ -1340,7 +1336,6 @@ void
 MLASCALL
 MlasNchwcPool(
     MLAS_POOLING_KIND PoolingKind,
-    size_t Dimensions,
     const int64_t* InputShape,
     const int64_t* KernelShape,
     const int64_t* DilationShape,
@@ -1360,8 +1355,6 @@ Routine Description:
 Arguments:
 
     PoolingKind - Supplies the kind of pooling operation to perform.
-
-    Dimensions - Supplies the number of dimensions.
 
     InputShape - Supplies the shape of the input tensor.
 
@@ -1403,7 +1396,7 @@ Return Value:
     // Capture the generic shape parameters to the work block.
     //
 
-    MlasNchwcPrepareWorkBlock(&WorkBlock, Dimensions, InputShape, KernelShape,
+    MlasNchwcPrepareWorkBlock(&WorkBlock, InputShape, KernelShape,
         DilationShape, Padding, StrideShape, OutputShape);
 
     //
@@ -1413,6 +1406,109 @@ Return Value:
     WorkBlock.tids = MlasGetMaximumThreadCount(ThreadPool);
 
     MlasExecuteThreaded(MlasNchwcThreaded<MLAS_NCHWC_POOL_ALGORITHM>, &WorkBlock, WorkBlock.tids, ThreadPool);
+}
+
+void
+MLASCALL
+MlasNchwcUpsample(
+    const int64_t* InputShape,
+    const int64_t* Scales,
+    const float* Input,
+    float* Output
+    )
+/*++
+
+Routine Description:
+
+    This routine implements the NCHWc upsample nearest operation.
+
+Arguments:
+
+    InputShape - Supplies the shape of the input tensor.
+
+    Scales - Supplies the shape of the spatial scaling.
+
+    Input - Supplies the input tensor.
+
+    Output - Supplies the output tensor.
+
+Return Value:
+
+    None.
+
+--*/
+{
+    const size_t BlockSize = MlasNchwcGetBlockSize();
+
+    const size_t BatchCount = size_t(InputShape[0]);
+    const size_t ChannelCount = size_t(InputShape[1]);
+    const size_t InputHeight = size_t(InputShape[2]);
+    const size_t InputWidth = size_t(InputShape[3]);
+
+    const size_t TotalInputHeight = BatchCount * ChannelCount * InputHeight;
+
+    const size_t ScaleHeight = size_t(Scales[0]);
+    const size_t ScaleWidth = size_t(Scales[1]);
+
+    const size_t OutputWidth = InputWidth * ScaleWidth;
+
+    //
+    // Iterate over each line of the input tensor.
+    //
+
+    for (size_t h = 0; h < TotalInputHeight; h += BlockSize) {
+
+        float* OutputBaseRow = Output;
+
+        //
+        // Scale the input tensor across the width dimension.
+        //
+
+        for (size_t w = 0; w < InputWidth; w++) {
+
+            if (BlockSize == 16) {
+
+                MLAS_FLOAT32X4 v0 = MlasLoadFloat32x4(Input);
+                MLAS_FLOAT32X4 v1 = MlasLoadFloat32x4(Input + 4);
+                MLAS_FLOAT32X4 v2 = MlasLoadFloat32x4(Input + 8);
+                MLAS_FLOAT32X4 v3 = MlasLoadFloat32x4(Input + 12);
+
+                for (size_t sw = 0; sw < ScaleWidth; sw++) {
+
+                    MlasStoreFloat32x4(Output, v0);
+                    MlasStoreFloat32x4(Output + 4, v1);
+                    MlasStoreFloat32x4(Output + 8, v2);
+                    MlasStoreFloat32x4(Output + 12, v3);
+
+                    Output += BlockSize;
+                }
+
+            } else {
+
+                MLAS_FLOAT32X4 v0 = MlasLoadFloat32x4(Input);
+                MLAS_FLOAT32X4 v1 = MlasLoadFloat32x4(Input + 4);
+
+                for (size_t sw = 0; sw < ScaleWidth; sw++) {
+
+                    MlasStoreFloat32x4(Output, v0);
+                    MlasStoreFloat32x4(Output + 4, v1);
+
+                    Output += BlockSize;
+                }
+            }
+
+            Input += BlockSize;
+        }
+
+        //
+        // Scale the input tensor across the height dimension by duplicating
+        // the first output line.
+        //
+
+        for (size_t sh = 1; sh < ScaleHeight; sh++) {
+            Output = std::copy_n(OutputBaseRow, OutputWidth * BlockSize, Output);
+        }
+    }
 }
 
 #if !defined(MLAS_TARGET_AMD64)