**Description**: Extend the NCHWc layout optimizer to handle Resize(mode=nearest) and Mul.

**Motivation and Context**
Add support for trivial Upsample and Resize ops using nearest mode upsampling with integer scaling in the NCHWc domain. This is meant to handle UNet and YOLOV3 models that have an UpsampleNearest(2x2) node. This avoids unnecessary tensor reordering around the upsampling.

Also reuse the existing Add/Sum support to keep tensors in NCHWc when both inputs of a Mul node are already in NCHWc format.

One internal model using Upsample and Mul went from 27ms->21ms per inference. tinyyolov3 from the model zoo went from 16ms->14ms.